+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'chat'
+rootProject.name = 'gateway'
