+++ src/main/java/com/nemo/gateway/GatewayApplication.java
@@ -1,13 +1,13 @@
-package com.nemo.chat;
+package com.nemo.gateway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ChatApplication {
+public class GatewayApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(ChatApplication.class, args);
+        SpringApplication.run(GatewayApplication.class, args);
     }
 
 }
