+++ src/main/java/com/nemo/gateway/HelloController.java
@@ -1,4 +1,4 @@
-package com.nemo.chat;
+package com.nemo.gateway;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
