+++ src/main/java/com/nemo/gateway/WebConfig.java
@@ -1,4 +1,4 @@
-package com.nemo.chat;
+package com.nemo.gateway;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
