+++ src/test/java/com/nemo/gateway/GatewayApplicationTests.java
@@ -1,10 +1,10 @@
-package com.nemo.chat;
+package com.nemo.gateway;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ChatApplicationTests {
+class GatewayApplicationTests {
 
     @Test
     void contextLoads() {
